Key pass update logic from Pass#last_update
`Pass#last_update` is effectively the same thing (it runs through the
generator and looks at the `updated_at` value), but by hardcoding that
again here we can't overwrite the logic in our own `Pass` instances.
This is helpful to have in the case of complex logic (i.e., not tied to
a singular model) in a `Pass` that determines whether or not a user's
pass should be updated or not.